add categories complete and ready for testing
