+++ TabloidFullStack/TabloidFullStack/client/tabloid/src/components/Categories/CategoryForm.js
@@ -1,30 +1,49 @@
 import React, { useState } from 'react';
+import { addCategory } from '../../Managers/CategoryManager';
+import { useNavigate } from 'react-router';
 
-//export here does not work, not sure why, so exporting at bottom of file
+
 const CategoryForm = ({ onSave }) => {
   const [newCategoryName, setNewCategoryName] = useState('');
 
-  const handleSave = (e) => {
-    onSave(newCategoryName);
-    setNewCategoryName('');
+  const navigate = useNavigate()
+
+  const handleSave = async (e) => {
+    e.preventDefault(); 
+    try {
+      const newCategory = {
+        name: newCategoryName,
+      };
+      const response = await addCategory(newCategory);
+      if (response.ok) {
+        onSave(); 
+        setNewCategoryName(''); 
+
+      } else {
+        console.error('Failed to add category:', response.statusText);
+      }
+    } catch (error) {
+      console.error('Error adding category:', error);
+    }
+    navigate('/categories');
   };
 
   return (
-    <form onSubmit={handleSave}>
-      <label>
-        Category Name:
-        <input
-          type="text"
-          value={newCategoryName}
-          onChange={(e) => setNewCategoryName(e.target.value)}
-        />
-      </label>
-      <button type="submit">Save</button>
-    </form>
+    <div className='category-form'>
+      <h1>Create a new Category</h1>
+      <form onSubmit={handleSave}>
+        <label>
+          Category Name:
+          <input type="text" value={newCategoryName} onChange={(e) => setNewCategoryName(e.target.value)} />
+        </label>
+        <button type="submit">Save</button>
+      </form>
+    </div>
   );
 };
 
 export default CategoryForm;
+
