+++ TabloidFullStack/TabloidFullStack/client/tabloid/src/components/Categories/CategoryList.js
@@ -1,33 +1,25 @@
+//Imports go here
 import React, { useEffect, useState } from "react";
 import { addCategory, getAllCategories } from "../../Managers/CategoryManager";
 import CategoryForm from "./CategoryForm";
 
+
+/*defining CategoryList, the main function of CategoryList.js that gets exported as a compenent for ApplicationViews.js. For some reason
+I cannot get the export to work when I first define CategoryList, so I export it at the bottom of this file */
 const CategoryList = () => {
+  //Uses state: categories holds the result of my GetAllCategories. See const getCategories below, which uses setCategories to do this
   const [categories, setCategories] = useState([]);
+  //CategoryForm is like a lightswitch; when true the input for a new category is seen. When false (as is default state below), input is turned off
   const [categoryForm, setCategoryForm] = useState(false);
 
+  //As noted above, getCategories first does a fetch (getAllCategories) and then uses setCategories from Line 11 to update "categories" to fetch result
   const getCategories = () => {
     getAllCategories().then((allCategories) => setCategories(allCategories));
   };
 
-  const handleSave = async (newCategoryName) => {
-    try {
-      const newCategory = {
-        name: newCategoryName,
-      };
-      const response = await addCategory(newCategory);
-      if (response.ok) {
-        getCategories();
-      } else {
-        console.error("Failed to add category:", response.statusText);
-      }
 
-      setCategoryForm(false);
-    } catch (error) {
-      console.error("Error adding category:", error);
-    }
-  };
 
+  //do a new fetch request to get all Categories, which may or may not now be updated with a new one after saving (see handleSave above)
   useEffect(() => {
     getCategories();
   }, []);
@@ -50,8 +42,10 @@ const CategoryList = () => {
           ))}
         </tbody>
       </table>
-      <button onClick={() => setCategoryForm(true)}>Add</button>
-      {categoryForm && <CategoryForm onSave={handleSave} />}
+      {/*the Add button below sets setCategoryFrom to true, which makes the input appear */}
+      <a href="http://localhost:3000/categories/form"><button>Create Category</button></a>
+      {/* When save is clicked, handleSave runs */}
+      
     </div>
   );
 };
