+++ TabloidFullStack/TabloidFullStack/client/tabloid/src/index.css
@@ -16,3 +16,14 @@ code {
 td, th{
   padding-left: 3rem;
 }
+
+.category-form{
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  padding: 1rem;
+}
+
+.category-form-input{
+  margin: 1rem;
+}
