+++ package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "react-router-dom": "^6.22.0"
+  }
+}
